Added support for empty ensemble models
Helper visitors (such as field resolvers) should not raise
markup exceptions related to empty model bodies.
For example, it would be justified to raise a missing element exception
when a MiningModel element is missing a primary Segmentation child element,
but not when the Segmentation element itself is missing secondary
Segment child element(s).